OMAP:DSS: Allow specifying which DSS callbacks to get
Add a bitmask to the callback specification.  This allows limiting
which callback to get from the DSS.

The callback function must return a new mask that is and-ed to
the existing mask.  If the mask becomes 0, no more callbacks will
be scheduled for the settings that are tracked.

The next patch will optimize IRQ scheduling, so that the CPU is
woken up only for the requested callbacks.

Change-Id: Id9054c44311a7e99aa9996cc9831d72bd460c208
Signed-off-by: Lajos Molnar <molnar@ti.com>