OMAP:DSS: WA for HW issue in dss_apply_irq_handler where GO is still set
Sometimes when handling an interrupt for a manager, the manager is still
busy at the beginning of the interrupt handler.   Later it becomes idle, so
we unregister the interrupt.  This leaves the shadow_dirty flag in an
incorrect true state, and also misses the 'programmed' callback.

For now, we do not unregister the interrupt if any manager was busy at the
first read of the GO bits.  A better fix would be to keep the first read
busy state in the cache, so we do not operate on instantaneous reads of the
GO bit.

Change-Id: I26a2f5a390d9eb24a7ca3320da5cb713eebf6299
Signed-off-by: Lajos Molnar <molnar@ti.com>